iDLDDG/requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.2.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==4.1.0
blinker==1.6.2
Bottleneck==1.3.7
Brotli==1.0.9
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.1
cryptography==41.0.3
cycler==0.12.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==0.8.3
fair-esm==2.0.0
fastjsonschema==2.16.2
filelock==3.14.0
flatbuffers==20210226132247
fonttools==4.25.0
frozenlist==1.4.0
fsspec==2024.3.1
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.9.0
huggingface-hub==0.26.3
idna==3.4
importlib-metadata==7.0.1
importlib_resources==6.4.0
intel-openmp==2021.4.0
ipykernel==6.28.0
ipython==8.15.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==5.5.0
jupyter-events==0.6.3
jupyter-server==1.13.5
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
llvmlite==0.43.0
Markdown==3.4.1
MarkupSafe==2.1.3
matplotlib==3.8.4
matplotlib-inline==0.1.6
mistune==0.8.4
mkl==2021.4.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.4
munkres==1.1.4
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==5.5.0
nbformat==5.9.2
nest-asyncio==1.5.6
networkx==3.2.1
notebook==6.5.4
notebook_shim==0.2.3
numba==0.60.0
numexpr==2.8.4
numpy==1.22.3
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.0
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
pip==23.3.1
platformdirs==3.10.0
pooch==1.7.0
prometheus-client==0.14.1
prompt-toolkit==3.0.43
protobuf==3.14.0
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.15.1
PyJWT==2.4.0
pyOpenSSL==23.2.0
pyparsing==3.1.2
pyrsistent==0.20.0
PySocks==1.7.1
pyteomics==4.7.5
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2024.1
pywin32==305.1
pywinpty==2.0.10
PyYAML==6.0.1
pyzmq==24.0.1
regex==2024.11.6
requests==2.31.0
requests-oauthlib==1.3.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.7.2
safetensors==0.4.5
scikit-learn==1.2.2
scipy==1.12.0
seaborn==0.13.2
Send2Trash==1.8.2
sentencepiece==0.2.0
setuptools==68.2.2
sip==4.19.13
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.2.0
sympy==1.12
tbb==2021.12.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-addons==0.22.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.17.1
testpath==0.6.0
threadpoolctl==2.2.0
tinycss2==1.2.1
tokenizers==0.20.3
torch==2.3.0
tornado==6.3.3
tqdm==4.65.0
traitlets==5.7.1
transformers==4.46.3
typeguard==2.13.3
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.2.3
wheel==0.35.1
win-inet-pton==1.1.0
wrapt==1.12.1
xgboost==2.1.4
xlrd==1.1.0
yarl==1.9.3
zipp==3.17.0
